refactor: moved buttons to a common directory
