--- src/Components/Common/Button/Button.module.scss
@@ -1,5 +1,0 @@
-@use '../../../utils/styles/mixins' as *;
-
-.button {
-  @extend %button;
-}
