+++ src/Components/Common/Buttons/Button/Button.module.scss
@@ -1,0 +1,5 @@
+@use '../../../../utils/styles/mixins' as *;
+
+.button {
+  @extend %button;
+}
