+++ src/Components/Common/Buttons/LanguageButton/LanguageButton.module.scss
@@ -1,4 +1,4 @@
-@use '../../../utils/styles/variables' as *;
+@use '../../../../utils/styles/variables' as *;
 
 .languageButton {
   width: 70px;
