+++ src/Components/Header/Header.tsx
@@ -1,7 +1,7 @@
 import type { FC, ReactElement } from 'react';
 import classes from './Header.module.scss';
 import { Navigation } from '../Common/Navigation/Navigation';
-import { LanguageButton } from '../Common/LanguageButton/LanguageButton';
+import { LanguageButton } from '../Common/Buttons/LanguageButton/LanguageButton';
 
 export const Header: FC = (): ReactElement => {
   return (
